+++ addons/io_scene_gltf2/io/imp/gltf2_io_bufferview.py
